JLFP: Task-aware wss overhead charges.
The JLFP overhead model currently charges the same max-wss CPMD
to all tasks of a task set, even if each task has a different
wss. Instead, each task should be charged the max-wss CPMD, except
for the task with the largest wss. This task should be charged
a CPMD for the second-greatest wss.